ensure proper return type

+++ src/main.h
@@ -1404,7 +1404,7 @@ inline int64_t GetMaxStakeAge(CBlockIndex* pindex) {
     int nMaxStakeAge = nMinStakeAge * 2;
 
     // disallow stakes longer than 1 month.
-    return nMaxStakeAge >= 30 ? 30 : nMaxStakeAge;
+    return nMaxStakeAge >= 30 ? (int64_t)30 : (int64_t)nMaxStakeAge;
 }
 
 #endif
